Updates to SevenSegmentDecoder.vhd
Added the SevenSegmentDecoder updates, and looking at the modelsim graph and comapring it to the Lab 2 simulation we've submitted, there isn't a whole lot that is different. One difference I noticed with this change as well as the registers.vhd change is that the the simulation posted has the same results, aside from the fact that our Selector bit is inverted. Something to keep in mind.