Fixed binary_bcd.vhd!
Made changes to the two-state machine. I found that it was NOT necessary for reset to equal '0' for a state change to occur, as the previous code had an else if with regards to reset. I removed reset from the sequential process sensitivity list and the if statement's condition.

The only problem in all my simulations so far is the Selector bit being the opposite of what we found in Lab 2.